update favourite apps list, fix recent documents create

+++ setup_recentdocuments.sh
@@ -5,4 +5,8 @@ if [ ! -L $HOME/.kde4/share/apps/RecentDocuments ] && [ -d $HOME/.kde4/share/app
 fi
 
 [ ! -d $HOME/.kde4/share/apps ] && mkdir -p $HOME/.kde4/share/apps
-ln -sf $HOME/.local/share/RecentDocuments $HOME/.kde4/share/apps/RecentDocuments
+
+if [ ! -L $HOME/.kde4/share/apps/RecentDocuments ] || [ "$(readlink $HOME/.kde4/share/apps/RecentDocuments)" = "$HOME/.local/share/RecentDocuments" ]; then
+    rm -rf $HOME/.kde4/share/apps/RecentDocuments
+    ln -sf $HOME/.local/share/RecentDocuments $HOME/.kde4/share/apps/RecentDocuments
+fi
